\name{demogR-package}
\alias{demogR-package}
\alias{demogR}
\docType{package}
\title{
	Demographic analysis in R

}
\description{
\code{demogR} is a collection of functions to construct and analyze
age-structured population models.  For a list of functions, type:
\code{help(package="demogR")}.
}
\details{
\tabular{ll}{
Package: \tab demogR\cr
Type: \tab Package\cr
Version: \tab 0.5.0\cr
Date: \tab 2017-04-04\cr
License: \tab GPL version 2.0 or newer \cr
}

}
\author{ James Holland Jones\cr
	      Department of Anthropological Sciences\cr
	      Stanford University\cr

Maintainer: Hana Sevcikova <hanas@uw.edu>

}
\references{
Caswell, H. 2001. Matrix population models: Construction, analysis,
and interpretation. 2nd ed. Sunderland, MA: Sinauer. 

Coale, A.J. 1972. The growth and structure of human populations: A
mathematical investigation. Princeton: Princeton University Press. 

Jone, J.H. (2007). demogR: A Package for the Construction and Analysis of Age-structured Demographic Models in R. Journal of Statistical Software, 22(10), 1-28.

Keyfitz, N. 1977. Introduction to the mathematics of populations. 2nd
ed. Menlo Park: Addison-Wesley.

Keyfitz, N., and H. Caswell. 2005. Applied mathematical
demography. 3rd ed. New York: Springer. 

Leslie, P.H. 1945. On the use of matrices in certain population
mathematics. Biometrics 33:213-245. 

Preston, S.H., P. Heuveline, and F. Guillot. 2001. Demography:
Measuring and modeling population processes. Oxford: Blackwell. 

Tuljapurkar, S. 1990. Population dynamics in variable environments.
  Vol. 85, Lecture notes in biomathematics. Berlin: Springer-Veralg.
  }


\keyword{ package }
